make the webui built with mnist demo
Note: when using cmake to build qt project, the header file name needs to match the name of the cpp.

Otherwise, some linking error will appear. This seems to be a bug of cmake.

Before this change, the HttpServer object was defined in Swiftly.h, but implemented in HttpServer.cpp, this caused linking trouble of MNIST.

After changing the cpp to Swiftly.cpp, the issue went away.