fix xor model bug
the bug was because I forgot to clear derivative tensors. Those derivative operators are accumulative, the output tensors of those operators need to be cleared each iteration.

there is also a bug in the model->clearTensor function, where I should start enumeration from tensor zero, but instead I started with tensor 1, because this code was copied and pasted from the endMutateTensor function