fix multi-cpu ConvNet model training
the issue was due to a race condition when reshaping.

reshaping is currently carried out in the main thread, but

evaluation is carried out by work threads. There was no synchronization

before reshaping to wait for previous operator to finish evaluation.

Reshaping mechanism should be redesigned.

Sanity should pass if tensor is reshaped during initialization.

Reshape should be carried out by worker thread.

A possible solution is extending operator to accept a Tensor with reshaping

as an input.